+++ tests/test_cli_import_solution.py
@@ -585,6 +585,21 @@ class TestCliImportSolution(object):  # pylint: disable=too-many-public-methods
             assert not Database.get_collection().size()
             mock_file.assert_called_once_with('./solution-template.txt', 'r')
 
+    @pytest.mark.usefixtures('isfile_true')
+    def test_cli_import_solution_029(self, snippy):
+        """Import all solutions.
+
+        Try to import content with option --all. This is not supported for
+        import operation.
+        """
+
+        with mock.patch('snippy.content.migrate.open', mock.mock_open(), create=True) as mock_file:
+            cause = snippy.run(['snippy', 'import', '--all', '-f', './foo.yaml'])
+            assert cause == 'NOK: content category \'all\' is supported only with search or export operation'
+            assert not Database.get_collection().size()
+            mock_file.assert_not_called()
+
+
     @classmethod
     def teardown_class(cls):
         """Teardown class."""